consumer/app/requirements.txt
kafka-python
aiokafka
fastapi
uvicorn
confluent_kafka
asyncio
cassandra-driver
pyyaml
uuid